Suggestion on using a IP scanner to find the IP where it is(Home Assistant)
I have Made Changes to the 5th Point...
Here is the change I made...
After the first sentence of the 5th point i have inserted this:

In some cases, When the application is idling and cursor is blinking hit the "Enter" key and the HassIO screen should appear, thenat the "Homeassistant login:" appears type 'root' and hit enter, Now use any IP scanner for eg. "Advanced IP Scanner" then hit scan after installing the application then when homeassistant appears on the list of IPs right click and click on copy from dropdown menu and choose IP, now go to any browser(Chrome is recommended), and paste the IP address and add ":8123" in the end and hit enter to get to the Home Assistant Site.